Optimize DFM: parallel layers, compact rasters, contour/compare fixes
- Parallelize layer processing with Rayon (layers are independent)
- Use Vec<u8> for binary rasters instead of Vec<f64> (8x memory reduction);
  f64 intermediates live only inside the Gaussian blur computation
- Fix contour chaining to use VecDeque (O(1) front-extension vs O(n))
- Fix marching squares saddle point disambiguation by sampling center value
- Use designed geometry half-diagonal for feature-erased edge deviation
  fallback instead of the full raster diagonal
- Deduplicate default sigma/threshold constants across Rust and Python

+++ python/rosette/_core.pyi
@@ -994,7 +994,7 @@ class LayerPrediction:
     metrics: LayerMetrics | None
     violations: list[DfmViolation]
     has_raster: bool
-    raster_data: list[float] | None
+    raster_data: list[int] | None
     raster_width: int | None
     raster_height: int | None
     raster_origin: tuple[float, float] | None
